+++ CHANGELOG.md
@@ -24,7 +24,9 @@ Requires Panda system version >= 1.1.0
 
   * Changes in network protocol for new panda system version
   * Removed unnecessary public dependencies for libfranka
+  * More descriptive exception messages
   * CI: Run linter on examples
+  * Examples first move to an initial joint position
 
 ## 0.1.0 - 2017-09-15